Author post related work

+++ admin/includes/view_all_posts.php
@@ -27,9 +27,34 @@ if (isset($_POST['checkBoxArray'])) {
           die("Query Failed. ". mysqli_error($connection));
         }
         break;
+
+        case 'clone':
+        $query = "SELECT * FROM posts WHERE post_id = {$postValueId} ";
+        $select_posts = mysqli_query($connection, $query);
+        while ($row = mysqli_fetch_assoc($select_posts)) {
+            $post_author = $row['post_author'];
+            $post_title = $row['post_title'];
+            $post_category_id = $row['post_category_id'];
+            $post_status = $row['post_status'];
+            $post_image = $row['post_image'];
+            $post_tags = $row['post_tags'];
+            $post_comment_count = $row['post_comment_count'];
+            $post_date = $row['post_date'];
+            $post_content = $row['post_content'];
+        }
+
+        $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags, post_comment_count, post_status)";
+
+        $query .= " VALUES({$post_category_id}, '{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}', '{$post_tags}', '{$post_comment_count}', '{$post_status}')";
+
+        $clone_post_query = mysqli_query($connection, $query);
+        confirmquery($clone_post_query);
+
+
+
       
       default:
-        # code...
+
         break;
     }
 
@@ -44,6 +69,7 @@ if (isset($_POST['checkBoxArray'])) {
       <option value="">Select Options</option>
       <option value="published">Published</option>
       <option value="draft">Draft</option>
+      <option value="clone">Clone</option>
       <option value="delete">Delete</option>
     </select>
   </div>
@@ -64,14 +90,15 @@ if (isset($_POST['checkBoxArray'])) {
           <th scope="col">Tags</th>
           <th scope="col">Comments</th>
           <th scope="col">Date</th>
-          <th scope="col">View</th>
+          <th scope="col">View Post</th>
+          <th scope="col">Views</th>
           <th scope="col" colspan="2">Actions</th>
         </tr>
       </thead>
       <tbody>
         <?php
 
-        $query = "SELECT * FROM posts";
+        $query = "SELECT * FROM posts ORDER BY post_id DESC";
         $select_posts = mysqli_query($connection, $query);
 
         $id = 1;
@@ -85,6 +112,7 @@ if (isset($_POST['checkBoxArray'])) {
             $post_tags = $row['post_tags'];
             $post_comment_count = $row['post_comment_count'];
             $post_date = $row['post_date'];
+            $post_views_count = $row['post_views_count'];
             
 
             echo "<tr>";
@@ -115,8 +143,9 @@ if (isset($_POST['checkBoxArray'])) {
             echo "<td>{$post_comment_count}</td>";
             echo "<td>{$post_date}</td>";
             echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
+            echo "<td><a href='posts.php?reset={$post_id}'>{$post_views_count}</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onclick=\" javascript: return confirm('Are you sure you want to delete it?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
 
             echo "</tr>"; 
         }
@@ -126,12 +155,23 @@ if (isset($_POST['checkBoxArray'])) {
     </form>
     <?php 
 
+    // delete a post
     if (isset($_GET['delete'])) {
       $the_post_id = $_GET['delete'];
 
       $query = "DELETE FROM posts WHERE post_id={$the_post_id}";
       $delete_query = mysqli_query($connection, $query);
       confirmquery($delete_query);
+      header("Location: posts.php");
+    }
+
+    // reset views count
+    if (isset($_GET['reset'])) {
+      $the_post_id = $_GET['reset'];
+
+      $query = "UPDATE posts SET post_views_count = 0  WHERE post_id= " .mysqli_real_escape_string($connection, $_GET['reset']) . " ";
+      $reset_query = mysqli_query($connection, $query);
+      confirmquery($reset_query);
       header("Location: posts.php");
     }
